% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxcombo_size.R
\name{GSMC_design}
\alias{GSMC_design}
\title{Sample size and boundary for GSMC design}
\usage{
GSMC_design(
  FHweights,
  interim_ratio,
  error_spend,
  eps,
  p,
  b,
  tau,
  omega,
  lambda,
  lambda.trt,
  rho,
  gamma,
  beta,
  stoch = TRUE,
  range_ext = 200,
  time_ext_multiplier = 1.5,
  time_increment = 0.01
)
}
\arguments{
\item{FHweights}{a list of pairs of pairs of parameters (rho, gamma) for the Fleming Harrington weighted log-rank test:\eqn{W(t)=S^\rho(t^-)(1-S(t^-))^\gamma}. The first one will provide the information fraction we used to sign the stopping time. Now we only accommodate the surrogate information fraction, i.e., rho = 0 and gamma = 0.}

\item{interim_ratio}{a vector of ratios between 0 and 1 that signing the stopping time following the first type of information fraction in FHweights. Note that the last  value must be 1, otherwise, it will automatically append a 1 for the final stage.}

\item{error_spend}{cumulative errors spent at each stage (including the final one). Must be of the same length as interim_ratio or one element shorter if the last value of interim_ratio is less than 1. The last element f error_spend should be type I error \code{alpha}.}

\item{eps}{the change point, before which, the hazard ratio is 1, and after which, the hazard ratio is theta}

\item{p}{treatment assignment probability.}

\item{b}{the number of subintervals per time unit.}

\item{tau}{the end of the follow-up time  in the study. Note that this is identical to \eqn{T+\tau} in the paper from Hasegawa (2014).}

\item{omega}{the minimum follow-up time for all the patients.  Note that Hasegawa(2014) assumes that the accrual is uniform between time 0 and R, and there does not exist any censoring except for the administrative censoring at the ending time \eqn{\tau}. Thus this value omega is equivalent to \code{tau-R}. Through our simulation tests, we found that this function is quite robust to violations of these assumptions: dropouts, different censoring rates for two  arms, and changing accrual rates.}

\item{lambda}{the hazard for the control group.}

\item{lambda.trt}{the hazard for the treatment group after time eps.}

\item{beta}{type II error, 1-power.}

\item{stoch}{stochastic prediction or not (exact prediction), the former is default.}

\item{range_ext}{the width to extend the range of sample size. If the predicted sampe size is not found in the range, try a wider \code{range_ext}. It will automatically search for the likely range, but still possibly miss the best one, Its default value is 200.}

\item{time_ext_multiplier}{the compute the  time window for the possible stopping time points by multiplying it with the total expected follow-up time tau. The default is 1.5, so that when tau = 18, the longest time we consider would be \eqn{18*1.5=27} months.}

\item{time_increment}{time increments to compute the predicted stopping time points, the finer the more accurate.}
}
\description{
Compute predicted sample size and boundaries for group sequential designs of max-combo tests.
}
