% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{readXlsx}
\alias{readXlsx}
\title{Read an Excel file}
\usage{
readXlsx(filename, sheetname)
}
\arguments{
\item{filename}{Character string indicating file path of Excel file}

\item{sheetname}{Character string indicating name of Excel sheet}
}
\description{
This is a wrapper around openxlsx::read.xlsx
}
