% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.R
\name{writeXlsx}
\alias{writeXlsx}
\title{Write an Excel file}
\usage{
writeXlsx(inputObject, filename)
}
\arguments{
\item{inputObject}{R object to write to Excel}

\item{filename}{Character string indicating file path of Excel file to write}
}
\description{
This is a wrapper around openxlsx::read.xlsx
}
