% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxcombo_size.R
\name{Maxcombo.beta.n}
\alias{Maxcombo.beta.n}
\alias{Maxcombo.beta.d}
\title{The power for a vector of sample sizes}
\usage{
Maxcombo.beta.n(Sigma1, mu1, z_alpha_vec, interim_vec, R, n_seq, n.rep = 5)

Maxcombo.beta.d(
  Sigma1,
  mu1,
  z_alpha_vec,
  interim_vec,
  R,
  d_seq,
  sum_D,
  n.rep = 5
)
}
\arguments{
\item{Sigma1}{the correlation matrix under the alternative hypothesis.}

\item{mu1}{the unit mu under the alternative hypothesis (the mean of the expectation of each subject scaled weighted log-rank test statistic, which can be approximated using the formula for \eqn{E^*} in Hasegawa 2014 paper. ).}

\item{z_alpha_vec}{same as the one exported from Maxcombo.bd, which is the boundaries for ordered test statistics, its order should be consistent to the rows and columns in \code{Sigma1}.}

\item{interim_vec}{the vector of the interims in each stages, not that it should be a repeat vector with same iterim values for all the test statitics at same stages.}

\item{R}{end of the enrollment time, which is identical to \code{R} defined in other functions like \code{\link{I.1}}.}

\item{n_seq}{the sequence of number of patients.}

\item{n.rep}{number of repeats to take the median for output}

\item{d_seq}{the sequence of number of expected events.}

\item{sum_D}{same as the exported value from \code{\link{sample.size_FH}}, the summed \eqn{D^*} in Hasegawa (2014).}
}
\description{
To obtain a spectrum of power for a vector of numbers of subjects (n) using \code{Maxcombo.beta.n} or events (d) using \code{Maxcombo.beta.d}.
}
\examples{
Sigma0_v<-rep(0.5,6)
Sigma0<-matrix(1, ncol=4,nrow=4)
Sigma0[upper.tri(Sigma0)]<- Sigma0_v
Sigma0[lower.tri(Sigma0)]<- t(Sigma0)[lower.tri(Sigma0)]
Sigma0
alpha_stage <- c(0.01,0.025) # The error you would like to spend at the interim stage
zz <- Maxcombo.bd(Sigma0 = Sigma0,index=c(1,1,2,2),alpha_sp=alpha_stage)
zz$z_alpha # boundary value for each stage
zz$z_alpha_vec # boundary value for each test statistic corresponding 
to the index
}
\seealso{
\code{\link{Maxcombo.sz}}
}
\author{
Lili Wang
}
