% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maxcombo_size.R
\name{stoch_pred}
\alias{stoch_pred}
\alias{stoch_pred.cov}
\title{A stochastic-process way of prediction}
\usage{
stoch_pred(eps, p, b, tau, omega, lambda, theta, rho, gamma, R)

stoch_pred.cov(eps, p, b, tau, omega, lambda, theta, rho1, gamma1, rho2,
  gamma2, R)
}
\arguments{
\item{eps}{Delayed treatment effect time.}

\item{p}{Probability of treatment assignment.}

\item{b}{The number of subintervals at each time point.}

\item{omega}{The minimum follow-up time for all the patients.  Note that Hasegawa(2014) assumes that the accrual is uniform between time 0 and R, and there does not exist any censoring except for the administrative censoring at the ending time \eqn{\tau}. Thus this value omega is equivalent to \code{tau-R}. Through our simulation tests, we found that this function is quite robust to violations of these assumptions: dropouts, different cenosring rates for two  arms, and changing accrual rates.}

\item{lambda}{The hazard for the control group.}

\item{theta}{The hazard ratio after the delayed time \code{eps} for the treatment arm.}

\item{rho, rho1, rho2}{The first parameter for Fleming Harrington weighted log-rank test:\eqn{W(t)=S^\rho(t^-)(1-S(t^-))^\gamma}.}

\item{R}{The accrual period.}
}
\value{
\item{sum_D}{The mean expected event ratio, multiplied by \code{n}, the sample size, it is equal to the stochastically predicted number of events. }
\item{inf or covariance}{The information/variance or covariance (averaged for each subject) , should multiplied by \code{n}, which is the sample size to obtain the stochastically predicted information. }
\item{E.star}{The unit mean, correspoinding to \eqn{E^*} in Hasegawa(2014)}
\item{trt_vs_ctrl_N}{The ratio of the samples sizes between the two arms, treatment vs control, corresonding to the time vector \code{t_vec}.}
\item{t_vec}{The time sequence corresponding to \code{trt_vs_ctrl_N}.}
}
\description{
A stochastic-process way of prediction of the expected event counts, mean difference, and the information(variance) or the covariance
}
\references{
Hasegawa, T. (2014). Sample size determination for the weighted log‐rank test with the Fleming–Harrington class of weights in cancer vaccine studies. Pharmaceutical statistics, 13(2), 128-135.
}
\author{
Lili Wang
}
